tests: add composition tests

+++ src/utils/utils.c
@@ -30,7 +30,7 @@ void split_image_into_rgb_channels(const unsigned char *image,
 								   struct image_rgb channel_image, int width,
 								   int height) {
 
-	for (int i = 0; i < width * height; i++) {
+	for (size_t i = 0; i < (size_t)width * (size_t)height; i++) {
 		channel_image.red[i] = image[i * 3 + 0];
 		channel_image.green[i] = image[i * 3 + 1];
 		channel_image.blue[i] = image[i * 3 + 2];
@@ -41,7 +41,7 @@ void assemble_image_from_rgb_channels(unsigned char *image,
 									  struct image_rgb channel_image, int width,
 									  int height) {
 
-	for (int i = 0; i < width * height; i++) {
+	for (size_t i = 0; i < (size_t)width * (size_t)height; i++) {
 		image[i * 3 + 0] = channel_image.red[i];
 		image[i * 3 + 1] = channel_image.green[i];
 		image[i * 3 + 2] = channel_image.blue[i];
